+++ routes/modules/activity/activity_teacher.js
@@ -11,16 +11,19 @@ const db = new Dao();
 
 /** 德育活动-班主任专栏-添加 */
 router.post("/add", async (req, res) => {
-	let { headline, department, author, publisher, timecreate, isTop, content, picSrc, fileList, fileListSrc, checked, removeSrc } = req.body;
-	
+	let { headline, author, timecreate, isTop, content, picSrc, fileList, fileListSrc, checked, removeSrc } = req.body;
+	const token = req.body.Authorization || req.query.token || req.headers["authorization"];
+    const publisher = utils.getJwtCode(token).username;
+    const publisherDepartmentId = (await utils.getUserData(publisher)).departmentId;
+    const publisherDepartmentName = (await db.find("system_department", {id: publisherDepartmentId}))[0].name;
 	try { 
 		await utils.removeAssets(removeSrc);
 	} catch (e) { 
 		res.status(200).send({ code: 500, msg: "图片删除失败", result: null });
 	}
 
-	const insertStr = {	headline: headline,	department: department,	author: author,	publisher: publisher, timecreate: timecreate, isTop: isTop,	topTime: timecreate, content: content, picSrc: picSrc, fileList: fileList, fileListSrc: fileListSrc, checked: checked,
-		id: uuidv1(), views: 0
+	const insertStr = {	headline: headline,	author: author,	publisher: publisher, timecreate: timecreate, isTop: isTop,	topTime: timecreate, content: content, picSrc: picSrc, fileList: fileList, fileListSrc: fileListSrc, checked: checked,
+		id: uuidv1(), views: 0, publisherDepartmentId: publisherDepartmentId, publisherDepartmentName: publisherDepartmentName
 	}
 	db.insertOne("activity_teacher", insertStr).then((success) => {
 		res.status(200).send({ msg: "保存成功", code: 200, result: success });
@@ -29,23 +32,20 @@ router.post("/add", async (req, res) => {
 
 /** 德育活动-班主任专栏-分页查询 */
 router.post("/query", async (req, res) => {
-	let { pageNo, pageSize, headline, author, department, startTime, endTime } = req.body;
+	let { pageNo, pageSize, headline, author, startTime, endTime } = req.body;
 
 	/** 参数校验区域 */
 	const sortStr = { "isTop": -1, "createtime": -1, "topTime": -1 };
 	const limitStr = {};
 	const regHeadline = new RegExp(headline, "i");
 	const regAuthor = new RegExp(author, "i");
-	const regDepartment = new RegExp(department, "i");
 	const whereStr = {
 		"headline": { $regex: regHeadline },
 		"author": { $regex: regAuthor },
-		"department": { $regex: regDepartment },
 		"$and": [{"timecreate": { "$gte": startTime }}, {"timecreate": { "$lte": endTime }}]
 	};
 	if (!headline) delete whereStr.headline;
 	if (!author) delete whereStr.author;
-	if (!department) delete whereStr.publisher;
 	if (!startTime && !endTime) delete whereStr.$and;
 	
 	await Promise.all([db.findByPage("activity_teacher", whereStr, limitStr, sortStr, pageNo, pageSize), db.getTotal("activity_teacher")]).then((array) => {
@@ -86,6 +86,16 @@ router.post("/changeIsTop", (req, res) => {
 router.post("/del", async (req, res) => {
 	let { id, fileListSrc, picSrc } = req.body;
 	const delStr = { "id": id };
+	const token = req.body.Authorization || req.query.token || req.headers["authorization"];
+    const publisher = utils.getJwtCode(token).username;
+    const publisherDepartmentId = (await utils.getUserData(publisher)).departmentId;
+    const publisherDepartmentName = (await db.find("system_department", {id: publisherDepartmentId}))[0].name;
+    const notHasDelAuth = await utils.notHasDelAuth(token, id, "activity_teacher");
+
+    if (notHasDelAuth) {
+    	res.status(200).send({ msg: "删除失败，没有权限", code: 500, result: {} });
+    	return;
+    } 
 	db.deleteOne("activity_teacher", delStr).then(async (success) => {
 		try {
 			await utils.removeAssets([...fileListSrc, ...picSrc]);
@@ -118,7 +128,7 @@ router.post("/queryById", async (req, res) => {
 
 /** 德育活动-班主任专栏-编辑 */
 router.post("/edit", async (req, res) => {
-	let { id, headline, department, author, publisher, timecreate, isTop, content, picSrc, fileList, fileListSrc, checked, topTime, removeSrc } = req.body;
+	let { id, headline, author, timecreate, isTop, content, picSrc, fileList, fileListSrc, checked, topTime, removeSrc } = req.body;
 
 	try {
 		await utils.removeAssets(removeSrc);
@@ -127,12 +137,19 @@ router.post("/edit", async (req, res) => {
 	}
 
 	let whereStr = { "id": id };
+	const token = req.body.Authorization || req.query.token || req.headers["authorization"];
+    const publisher = utils.getJwtCode(token).username;
+    const publisherDepartmentId = (await utils.getUserData(publisher)).departmentId;
+    const publisherDepartmentName = (await db.find("system_department", {id: publisherDepartmentId}))[0].name;
+    const notHasDelAuth = await utils.notHasDelAuth(token, id, "activity_teacher");
 
+    if (notHasDelAuth) {
+    	res.status(200).send({ msg: "编辑失败，没有权限", code: 500, result: {} });
+    	return;
+    } 
 	let updateStr = { $set: {
 		"headline": headline,
-		"department": department,
-		"author": author,
-		"publisher": publisher,		
+		"author": author,		
 		"timecreate": timecreate,		
 		"isTop": isTop,
 		"content": content,
@@ -140,7 +157,10 @@ router.post("/edit", async (req, res) => {
 		"fileList": fileList,
 		"fileListSrc": fileListSrc,
 		"checked": checked,
-		"topTime": isTop ? topTime : timecreate
+		"topTime": isTop ? topTime : timecreate,
+		"publisher": publisher,
+		"publisherDepartmentId": publisherDepartmentId,
+		"publisherDepartmentName": publisherDepartmentName
 	}};
 
 	db.updateOne("activity_teacher", whereStr, updateStr).then((success) => {
